:recycle: move current gcn dir to gcn/pad_pattern
