+++ jaxchem/models/__init__.py
@@ -1,2 +1,2 @@
 from jaxchem.models.nn import graph_pooling, clipped_sigmoid  # noqa
-from jaxchem.models.gcn import GCNLayer, GCN, GCNPredicator  # noqa
+from jaxchem.models.gcn import PadGCNLayer, PadGCN, PadGCNPredicator # noqa
