+++ jaxchem/models/nn/__init__.py
@@ -1,2 +1,2 @@
 from jaxchem.models.nn.clipped_sigmoid import clipped_sigmoid  # noqa
-from jaxchem.models.nn.graph_pooling import graph_pooling  # noqa
+from jaxchem.models.nn.graph_pooling import pad_graph_pooling  # noqa
