+++ tests/models/gcn/pad_pattern/test_gcn.py
@@ -2,18 +2,18 @@ import unittest
 
 import haiku as hk
 import jax.random as jrandom
-from jaxchem.models import GCNLayer
+from jaxchem.models import PadGCN
 
 
 # params
 in_feats = 64
-out_feats = 32
+hidden_feats = [64, 32, 16]
 batch_size = 32
 max_node_size = 30
 
 
-class TestGCNlayer(unittest.TestCase):
-    """Test GCNLayer"""
+class TestPadGCN(unittest.TestCase):
+    """Test PadGCN"""
 
     def setup_method(self, method):
         self.key = hk.PRNGSequence(1234)
@@ -25,12 +25,12 @@ class TestGCNlayer(unittest.TestCase):
         return (batched_node_feats, batched_adj, True)
 
     def __forward(self, node_feats, adj, is_training):
-        model = GCNLayer(in_feats=in_feats, out_feats=out_feats)
+        model = PadGCN(in_feats=in_feats, hidden_feats=hidden_feats)
         return model(node_feats, adj, is_training)
 
     def test_forward_shape(self):
-        """Test output shape of GCNLayer"""
+        """Test output shape of PadGCN"""
         forward = hk.transform(self.__forward, apply_rng=True)
         params = forward.init(next(self.key), *self.input_data)
         preds = forward.apply(params, next(self.key), *self.input_data)
-        assert preds.shape == (batch_size, max_node_size, out_feats)
+        assert preds.shape == (batch_size, max_node_size, hidden_feats[-1])
